+++ austrakka/__init__.py
@@ -1,4 +1,4 @@
 # NOTE: do not change the name of this variable or the quote type,
 # doing so will break github actions.
-__version__ = "0.73.0"
+__version__ = "0.73.2"
 __prog_name__ = "AusTrakka"